model/requirements.txt
fastapi
uvicorn[standard]
langchain-groq
langchain-core
sentence-transformers
faiss-cpu
numpy
matplotlib
bert-score
pydantic
